ELEC-231 created the modules for debouncer, added the masking function to stm32_interrupt
