+++ libraries/STM32F0xx_StdPeriph_Driver/src/stm32f0xx_interrupt.c
@@ -92,3 +92,17 @@ StatusCode stm32f0xx_interrupt_exti_clear_pending(uint8_t line) {
   EXTI_ClearITPendingBit((uint32_t)0x01 << line);
   return STATUS_CODE_OK;
 }
+
+StatusCode stm32f0xx_interrupt_exti_mask_set(uint8_t line, bool masked) {
+  if (line >= 16) {
+    return status_code(STATUS_CODE_INVALID_ARGS);
+  }
+  
+  if (masked) {
+    EXTI->IMR &= ~((uint32_t)1 << line);
+  } else {
+    EXTI->IMR |= (uint32_t)1 << line;
+  }
+
+  return STATUS_CODE_OK;
+}
